+++ src/main.h
@@ -26,7 +26,5 @@ int F_isalpha(int c);
 int F_isdigit(int c);
 int F_ischar(int c);
 void farsi_fkey(cmdarg_T *cap);
-int arabic_shape(int c, int *ccp, int *c1p, int prev_c, int prev_c1,
-                 int next_c);
 /* vim: set ft=c : */
 #endif /* NEOVIM_MAIN_H */
